refac: moved all relevant files to social_media app
Still work in progress - need to fix configs.
