+++ apps/social_feeds/lib/social_feeds.ex
@@ -1,18 +1,25 @@
 defmodule SocialFeeds do
   @moduledoc """
-  Documentation for SocialFeeds.
+  The SocialFeeds app endpoint functions.
   """
 
+  alias SocialFeeds.Cache
+
   @doc """
-  Hello world.
+  Returns the next (upcoming) meetup_event.
+  If there is no upcoming event, returns the last (recent) meetup_event.
+
+  This function does not hit external API if there is an unexpired next_meetup_event in cache.
 
   ## Examples
 
-      iex> SocialFeeds.hello
-      :world
+      iex> get_next_meetup_event()
+      %SocialFeeds.Meetup.Event{}
 
   """
-  def hello do
-    :world
+  def get_next_meetup_event do
+    Cache.fetch(:next_meetup_event,
+                fn -> SocialFeeds.Meetup.ApiClient.get_next_event() end,
+                %{cache_ttl_in_msec: 600_000})
   end
 end
