+++ apps/social_feeds/lib/social_feeds/application.ex
@@ -6,15 +6,6 @@ defmodule SocialFeeds.Application do
   use Application
 
   def start(_type, _args) do
-    # List all child processes to be supervised
-    children = [
-      # Starts a worker by calling: SocialFeeds.Worker.start_link(arg)
-      # {SocialFeeds.Worker, arg},
-    ]
-
-    # See https://hexdocs.pm/elixir/Supervisor.html
-    # for other strategies and supported options
-    opts = [strategy: :one_for_one, name: SocialFeeds.Supervisor]
-    Supervisor.start_link(children, opts)
+    SocialFeeds.Supervisor.start_link()
   end
 end
