+++ apps/social_feeds/lib/social_feeds/meetup/event.ex
@@ -1,4 +1,4 @@
-defmodule MontrealElixir.SocialFeeds.Meetup.Event do
+defmodule SocialFeeds.Meetup.Event do
   @moduledoc """
   Struct keeping meetup event attributes.
   """
